+++ data-migrations/2025-02-27-0917-add-api-info-to-affiliations.sql
@@ -1,6 +1,3 @@
-
+# Add the API Target location to the affiliations table
 ALTER TABLE affiliations
-  ADD COLUMN apiTarget VARCHAR(255) NULL,
-  ADD COLUMN apiAuthTarget VARCHAR(255) NULL,
-  ADD COLUMN apiAuthClientId VARCHAR(255) NULL,
-  ADD COLUMN apiAuthClientSecret VARCHAR(255) NULL;
+  ADD COLUMN apiTarget VARCHAR(255) NULL;