+++ src/schemas/affiliation.ts
@@ -32,7 +32,9 @@ export const typeDefs = gql`
     "Whether or not this affiliation is a funder"
     funder: Boolean!
     "The categories the Affiliation belongs to"
-    types: [AffiliationType!]!
+    types: [AffiliationType!]
+    "Has an API that be used to search for project/award information"
+    apiTarget: String
   }
 
   "The provenance of an Affiliation record"
@@ -118,14 +120,8 @@ export const typeDefs = gql`
     feedbackEmails: [String!]
     "The properties of this object that are NOT editable. Determined by the record's provenance"
     uneditableProperties: [String!]!
-    "(Funder only) The API URL used to search for projects/awards"
+    "The API URL that can be used to search for project/award information"
     apiTarget: String
-    "(Funder only) The Authorization endpoint used to retrieve an access token for the API"
-    apiAuthTarget: String
-    "(Funder only) The Client ID used to retrieve an access token for the API"
-    apiAuthClientId: String
-    "(Funder only) The Client Secret used to retrieve an access token for the API"
-    apiAuthClientSecret: String
   }
 
   "A collection of errors related to the Affiliation"
@@ -151,10 +147,6 @@ export const typeDefs = gql`
     feedbackMessage: String
     feedbackEmails: String
     subHeaderLinks: String
-    apiTarget: String
-    apiAuthTarget: String
-    apiAuthClientId: String
-    apiAuthClientSecret: String
   }
 
   "Input for a hyperlink displayed in the sub-header of the UI for the afiliation's users"